bp4dbg: restructure for proper python module
