+++ source/utils/CMakeLists.txt
@@ -57,9 +57,8 @@ if(ADIOS2_HAVE_MPI)
   add_subdirectory(adios_iotest)
 endif()
 
-# bp4dbg Python tool
-install(PROGRAMS bp4dbg/bp4dbg.py bp4dbg/bp4dbg_data.py bp4dbg/bp4dbg_utils.py 
-        bp4dbg/bp4dbg_metadata.py bp4dbg/bp4dbg_idxtable.py
-        DESTINATION ${CMAKE_INSTALL_BINDIR}/bp4dbg)
+if(Python_Interpreter_FOUND)
+  add_subdirectory(bp4dbg)
+endif()
 
 install(PROGRAMS adios_deactivate_bp.sh DESTINATION ${CMAKE_INSTALL_BINDIR})
